Add files via upload
Hi everyone,
I am very excited to share a project that I have completed on Excel. The name of the project is " Sales Dashboard". The main objective of the project is to create a sales dashboard using Excel to analyze factors affecting the sales and profit,correlation between discount and profit and top customers who are making profit.

Key insights : 
-Profit has an inverse relation with Discount.As Discount increases profit decreases.

-Tamara chand is the customer who makes the highest profit.

-The states which have maximum number  of sales in California.

-Yearly profits are gained in the technology and office category.